Move AdvancedSearchFormMigration from v3.4.2 to v3.4.3
