+++ web/src/main/webapp/WEB-INF/classes/setup/sql/migrate/v343/AdvancedSearchFormMigration.java
@@ -21,7 +21,7 @@
  * Rome - Italy. email: geonetwork@osgeo.org
  */
 
-package v342;
+package v343;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.annotations.VisibleForTesting;
