+++ web/src/test/java/v343/AdvancedSearchFormMigrationTest.java
@@ -21,7 +21,7 @@
  * Rome - Italy. email: geonetwork@osgeo.org
  */
 
-package v342;
+package v343;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -33,6 +33,7 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
+import v343.AdvancedSearchFormMigration;
 
 import java.io.IOException;
 import java.sql.Connection;
